MA-21494-2 Store the V4L2 OUTPUT device when enumerating video nodes.
The getVidoNode function will be called twice
during the initialization of DeviceAsWebcam.
The first is to check if there is a VIDEO_OUTPUT
device to determine whether the service should be started.
The second is to get the VIDEO_OUTPUT device
for use during the startService initialization.

We can store the output device if found at the first check,
then we can get the node directly instead of enumerating again
in later use.
This can help to save time when the service is running on some
platforms with many video nodes(and some may take much time to open),
which may avoid the setCurrentUsbFunctions timeout.

Signed-off-by: Anle Pan <anle.pan@nxp.com>
Change-Id: I0a39b35f848a26d970fc2a507f04766b728ddbe1